requirements.txt
alembic==0.9.10
arrow==0.12.1
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dateparser==0.7.0
flake8==3.5.0
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.8.1
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
moment==0.8.2
oauth2client==4.1.2
oauthlib==2.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
regex==2018.2.21
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
slackclient==1.2.1
SQLAlchemy==1.2.9
times==0.7
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.22
websocket-client==0.47.0
Werkzeug==0.14.1
